Updating Bed routes and tests.

+++ app/beds/bed_routes.py
@@ -1,15 +1,17 @@
+from typing import Annotated
 from uuid import UUID
 
-from fastapi import APIRouter, Depends, HTTPException
+from fastapi import APIRouter, Depends, HTTPException, Security
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from app.auth.auth import get_current_active_user
+from app.auth.auth import get_current_user
+from app.auth.auth_schemas import TokenData
 from app.beds.bed_crud import BedCRUD
 from app.beds.bed_models import Bed
 from app.beds.bed_schemas import BedCreate, BedList, BedRead, BedUpdate
+from app.core.auth.scopes_manager import ScopesManager
 from app.gardens.garden_crud import GardenCRUD
 from app.logging import get_logger, log_handler
-from app.users.user_models import User
 from app.utils.database import get_db
 
 logger = get_logger(__name__)
@@ -17,11 +19,29 @@ logger = get_logger(__name__)
 bed_router = APIRouter(prefix="/beds")
 
 
+def check_access(current_user: TokenData, garden_user: UUID, scope: str):
+
+    access_any = ScopesManager.has_scope(
+        user_scopes=current_user.scopes, required_scope=scope
+    )
+    is_owner = ScopesManager.is_owner(
+        user_scopes=current_user.scopes,
+        required_scope=f"{scope}:own",
+        user_id=current_user.id,
+        entity_owner=garden_user,
+    )
+
+    if not access_any and not is_owner:
+        raise HTTPException(status_code=403, detail="Forbidden")
+
+
 @bed_router.post("", response_model=Bed)
 async def create_bed(
     bed: BedCreate,
+    current_user: Annotated[
+        TokenData, Security(get_current_user, scopes=["ga:up", "ga:up:own"])
+    ],
     session: AsyncSession = Depends(get_db),
-    current_user: User = Depends(get_current_active_user),
 ) -> Bed:
     """
     Route to create a new bed in casbin.
@@ -31,6 +51,10 @@ async def create_bed(
     :param current_user: The User accessing the route
     :return: Bed; beds.bed_models.Bed
     """
+
+    garden = await GardenCRUD.get_garden(session=session, garden_id=bed.garden_id)
+    if not garden:
+        raise HTTPException(status_code=404, detail="Garden not found")
 
     new_bed = await BedCRUD.create_bed(
         bed=bed,
@@ -55,10 +79,12 @@ async def create_bed(
 @bed_router.get("/garden/{garden_id}", response_model=list[BedList])
 async def read_beds(
     garden_id: UUID,
+    current_user: Annotated[
+        TokenData, Security(get_current_user, scopes=["ga:re", "ga:re:own"])
+    ],
     skip: int = 0,
     limit: int = 100,
     session: AsyncSession = Depends(get_db),
-    current_user: User = Depends(get_current_active_user),
 ) -> list[BedList]:
     """
     Route to get all bed routes associated with a given garden.
@@ -75,6 +101,8 @@ async def read_beds(
     if not garden:
         raise HTTPException(status_code=404, detail="Garden not found")
 
+    check_access(current_user=current_user, garden_user=garden.user_id, scope="ga:re")
+
     beds = await BedCRUD.get_beds(
         garden_id=garden_id,
         session=session,
@@ -87,13 +115,14 @@ async def read_beds(
 @bed_router.get("/{bed_id}", response_model=BedRead)
 async def read_bed(
     bed_id: UUID,
+    current_user: Annotated[
+        TokenData, Security(get_current_user, scopes=["ga:re", "ga:re:own"])
+    ],
     session: AsyncSession = Depends(get_db),
-    current_user: User = Depends(get_current_active_user),
 ) -> BedRead:
     """
     Route to get a bed route associated with a given garden and bed id.
 
-    :param garden_id: Garden UUID
     :param bed_id: Bed UUID
     :param session: SQLAlchemy asyncio AsyncSession
     :param current_user: User
@@ -104,6 +133,12 @@ async def read_bed(
     if not bed:
         raise HTTPException(status_code=404, detail="Bed not found")
 
+    garden = await GardenCRUD.get_garden(session=session, garden_id=bed.garden_id)
+    if not garden:
+        raise HTTPException(status_code=404, detail="Garden not found")
+
+    check_access(current_user=current_user, garden_user=garden.user_id, scope="ga:re")
+
     return bed
 
 
@@ -111,8 +146,10 @@ async def read_bed(
 async def update_bed(
     bed_id: UUID,
     bed_update: BedUpdate,
+    current_user: Annotated[
+        TokenData, Security(get_current_user, scopes=["ga:up", "ga:up:own"])
+    ],
     session: AsyncSession = Depends(get_db),
-    current_user: User = Depends(get_current_active_user),
 ) -> Bed | None:
     """
     Route to update a bed
@@ -127,6 +164,12 @@ async def update_bed(
     bed = await BedCRUD.get_bed(session=session, bed_id=bed_id)
     if not bed:
         raise HTTPException(status_code=404, detail="Bed not found")
+
+    garden = await GardenCRUD.get_garden(session=session, garden_id=bed.garden_id)
+    if not garden:
+        raise HTTPException(status_code=404, detail="Garden not found")
+
+    check_access(current_user=current_user, garden_user=garden.user_id, scope="ga:up")
 
     updated_bed = await BedCRUD.update_bed(
         session=session,
@@ -152,8 +195,10 @@ async def update_bed(
 @bed_router.delete("/{bed_id}")
 async def delete_bed(
     bed_id: UUID,
+    current_user: Annotated[
+        TokenData, Security(get_current_user, scopes=["ga:de", "ga:de:own"])
+    ],
     session: AsyncSession = Depends(get_db),
-    current_user: User = Depends(get_current_active_user),
 ) -> dict:
     """
     Route to delete a bed
@@ -167,6 +212,12 @@ async def delete_bed(
     bed = await BedCRUD.get_bed(session=session, bed_id=bed_id)
     if not bed:
         raise HTTPException(status_code=404, detail="Bed not found")
+
+    garden = await GardenCRUD.get_garden(session=session, garden_id=bed.garden_id)
+    if not garden:
+        raise HTTPException(status_code=404, detail="Garden not found")
+
+    check_access(current_user=current_user, garden_user=garden.user_id, scope="ga:de")
 
     if not await BedCRUD.delete_bed(session, bed_id):
         raise HTTPException(status_code=404, detail="Bed not found")
